requirements.txt
pydantic==2.11.7
pyyaml==6.0.2